refactor(SiteCharts): extract common functions to utils
